face-finder-skeleton/backend/requirements.txt
# --- API core ---
fastapi==0.103.2
uvicorn[standard]==0.22.0
python-multipart==0.0.9
pydantic==1.10.13
starlette==0.27.0

# --- Auth & DB ---
sqlmodel==0.0.21
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# --- Image & ML ---
numpy==1.26.4
opencv-python==4.10.0.84
insightface==0.7.3
onnxruntime==1.19.0

# --- Utils ---
typing-extensions>=4.7
